Fixing tests, moving example file to example folder
