+++ src/components/sideNavigation/tests/SideNavigation.spec.tsx
@@ -1,0 +1,13 @@
+import { shallow, } from 'enzyme';
+import * as React from 'react';
+import { SideNavigation } from '../SideNavigation';
+
+describe('<SideNavigation />', () => {
+  it('should render without errors', () => {
+    expect(() => {
+      shallow(
+        <SideNavigation />
+      );
+    }).not.toThrow();
+  });
+});
