+++ src/components/sideNavigation/tests/SideNavigationLoadingHeader.spec.tsx
@@ -1,0 +1,12 @@
+import { shallow } from 'enzyme';
+import { SideNavigationLoadingHeader } from '../SideNavigationLoadingHeader';
+
+describe('<SideNavigationLoadingHeader />', () => {
+  it('should render without errors', () => {
+    expect(() => {
+      shallow(
+        <SideNavigationLoadingHeader />
+      );
+    }).not.toThrow();
+  });
+});
