+++ src/components/sideNavigation/tests/SideNavigationMenuSection.spec.tsx
@@ -1,0 +1,12 @@
+import { shallow } from 'enzyme';
+import { SideNavigationMenuSection } from '../SideNavigationMenuSection';
+
+describe('<SideNavigationMenuSection />', () => {
+  it('should render without errors', () => {
+    expect(() => {
+      shallow(
+        <SideNavigationMenuSection />
+      );
+    }).not.toThrow();
+  });
+});
